pyproject.toml
[tool.poetry]
name = "exante_data"
version = "0.1.0"
description = ""
authors = ["DanielsRepo <mcdancher18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.1"
beautifulsoup4 = "^4.12.3"
django-cors-headers = "^4.3.1" 
djangorestframework = "^3.14.0"
Django = "^5.0.2"
django-storages ="^1.12"
boto3 = "^1.34.54"
whitenoise = "^6.0.0"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"
python-http-client = "^3.3.7"
sendgrid = "^6.11.0"
sqlparse = "^0.4.4"
starkbank-ecdsa = "^2.2.0"
asgiref = "^3.7.2"
django-ordered-model = "^3.4.3"
psycopg2-binary = "^2.9.3"
dj-database-url = "^2.1.0"
django-admin-list-filter-dropdown = "^1.0.3"
django-admin-rangefilter = "^0.12.1"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
djoser = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
